Moved root container to application container and removed root container
