+++ src/screens/application/screen/applicationScreen.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { Component, Fragment } from 'react';
 import { IntlProvider } from 'react-intl';
 import { StatusBar, Platform, View } from 'react-native';
 import EStyleSheet from 'react-native-extended-stylesheet';
@@ -6,10 +6,15 @@ import { ReduxNavigation } from '../../../navigation/reduxNavigation';
 import { flattenMessages } from '../../../utils/flattenMessages';
 import messages from '../../../config/locales';
 
+// Services
+import { toastNotification as toastNotificationAction } from '../../../redux/actions/uiAction';
+
 // Components
 import { NoInternetConnection } from '../../../components/basicUIElements';
 import { ToastNotification } from '../../../components/toastNotification';
-import { toastNotification as toastNotificationAction } from '../../../redux/actions/uiAction';
+import { Modal } from '../../../components';
+import { PinCode } from '../../pinCode';
+import PostButtonForAndroid from '../../../components/postButton/view/postButtonsForAndroid';
 
 // Themes (Styles)
 import darkTheme from '../../../themes/darkTheme';
@@ -42,7 +47,14 @@ class ApplicationScreen extends Component {
   };
 
   render() {
-    const { isConnected, isDarkTheme, locale, toastNotification, isReady } = this.props;
+    const {
+      isConnected,
+      isDarkTheme,
+      locale,
+      toastNotification,
+      isReady,
+      isPinCodeReqiure,
+    } = this.props;
     const { isShowToastNotification } = this.state;
     const barStyle = isDarkTheme ? 'light-content' : 'dark-content';
     const barColor = isDarkTheme ? '#1e2835' : '#fff';
@@ -55,15 +67,21 @@ class ApplicationScreen extends Component {
           <StatusBar barStyle={barStyle} backgroundColor={barColor} />
         )}
 
-        {!isConnected && (
-          <IntlProvider locale={locale} messages={flattenMessages(messages[locale])}>
-            <NoInternetConnection />
-          </IntlProvider>
-        )}
-
         <IntlProvider locale={locale} messages={flattenMessages(messages[locale])}>
-          <ReduxNavigation />
+          <Fragment>
+            {!isConnected && <NoInternetConnection />}
+            <ReduxNavigation />
+            <Modal
+              isOpen={isPinCodeReqiure}
+              isFullScreen
+              swipeToClose={false}
+              backButtonClose={false}
+            >
+              <PinCode setWrappedComponentState={this._setWrappedComponentState} />
+            </Modal>
+          </Fragment>
         </IntlProvider>
+        {Platform.OS === 'android' && <PostButtonForAndroid />}
 
         {isShowToastNotification && (
           <ToastNotification
