+++ src/screens/index.js
@@ -12,7 +12,6 @@ import { Post } from './post';
 import { Profile } from './profile';
 import { Settings } from './settings';
 import { Voters } from './voters';
-import RootComponent from './root';
 import SteemConnect from './steem-connect/steemConnect';
 import { SearchResult } from './searchResult';
 import Transfer from './transfer';
@@ -30,7 +29,6 @@ export {
   PinCode,
   Post,
   Profile,
-  RootComponent,
   Settings,
   SteemConnect,
   Voters,
