--- src/screens/root/index.js
@@ -1,3 +1,0 @@
-import RootContainer from './container/rootContainer';
-
-export default RootContainer;
